feat: add employee management, housekeeping workflow and SePay webhook integration
- implement employee management module
  + create/update employee
  + reset employee password
  + role & permission handling

- implement housekeeping management module
  + create/update housekeeping tasks
  + task status tracking
  + room cleaning workflow

- integrate SePay payment gateway
  + checkout API integration
  + webhook callback processing
  + automatic payment status synchronization

- refactor related booking, payment and user flows
- add DTOs, repositories, services and controllers

+++ src/main/java/com/hotelmanagement/backend/dto/response/PromotionResponse.java
@@ -32,7 +32,7 @@ public class PromotionResponse {
     int quotaUsed;
     int quotaTotal;
     DiscountScope scope;
-    int minTotal;
-    int maxDiscountAmount;
+    BigDecimal minTotal;
+    BigDecimal maxDiscountAmount;
     boolean autoApplied;
 }
